# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Payment Sandbox API",
    "version": "1.0.0",
    "description": "HTTP API for deterministic payment integration testing."
  },
  "servers": [{"url": "http://127.0.0.1:8080"}],
...
  "paths": {
...
    "/admin/scenarios": {
      "post": {"tags": ["Administration"], "summary": "Create a scenario", "operationId": "createScenario", "security": [{"AdminBearerAuth": []}], "requestBody": {"$ref": "#/components/requestBodies/Scenario"}, "responses": {"201": {"$ref": "#/components/responses/Scenario"}, "400": {"$ref": "#/components/responses/InvalidRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "409": {"$ref": "#/components/responses/Conflict"}}}
    },
...
    "/admin/scenarios/{scenarioId}": {
      "parameters": [{"$ref": "#/components/parameters/ScenarioId"}],
      "get": {"tags": ["Administration"], "summary": "Inspect a scenario", "operationId": "getScenario", "security": [{"AdminBearerAuth": []}], "responses": {"200": {"$ref": "#/components/responses/Scenario"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}}}
    },
    "/admin/scenarios/{scenarioId}/execute": {
      "parameters": [{"$ref": "#/components/parameters/ScenarioId"}],
      "post": {"tags": ["Administration"], "summary": "Execute a stored scenario", "operationId": "executeScenario", "security": [{"AdminBearerAuth": []}], "responses": {"200": {"$ref": "#/components/responses/ScenarioExecution"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}}}
    },
...
  },
  "components": {
    "securitySchemes": {"AdminBearerAuth": {"type": "http", "scheme": "bearer", "description": "The PAYMENT_SANDBOX_ADMIN_TOKEN value."}},
    "parameters": {
...
      "ScenarioId": {"name": "scenarioId", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1}}
    },
    "requestBodies": {
...
      "Scenario": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Scenario"}}}}
    },
    "schemas": {
      "ErrorResponse": {"type": "object", "required": ["error"], "properties": {"error": {"type": "object", "required": ["code", "message"], "properties": {"code": {"type": "string"}, "message": {"type": "string"}}}}},
...
      "Payment": {"type": "object", "required": ["id", "amount", "currency", "status", "authorized_amount", "captured_amount", "refunded_amount", "version"], "properties": {"id": {"type": "string"}, "amount": {"type": "integer", "format": "int64"}, "currency": {"type": "string"}, "status": {"type": "string"}, "authorized_amount": {"type": "integer", "format": "int64"}, "captured_amount": {"type": "integer", "format": "int64"}, "refunded_amount": {"type": "integer", "format": "int64"}, "version": {"type": "integer", "format": "int64"}}},
...
      "Scenario": {"type": "object", "required": ["id", "provider", "initial_virtual_time", "deterministic_configuration", "initial_payments", "commands"], "additionalProperties": false, "properties": {"id": {"type": "string", "minLength": 1}, "provider": {"type": "object", "required": ["id"], "additionalProperties": false, "properties": {"id": {"type": "string", "minLength": 1}, "profile": {"type": "string"}}}, "initial_virtual_time": {"type": "string", "format": "date-time"}, "deterministic_configuration": {"type": "object", "required": ["seed"], "additionalProperties": false, "properties": {"seed": {"type": "integer", "format": "int64", "minimum": 0}}}, "initial_payments": {"type": "array", "items": {"$ref": "#/components/schemas/Payment"}}, "commands": {"type": "array", "items": {"$ref": "#/components/schemas/ScenarioCommand"}}}},
      "ScenarioCommand": {"type": "object", "required": ["type"], "additionalProperties": false, "properties": {"type": {"type": "string"}, "payment_id": {"type": "string"}, "amount": {"type": "integer", "format": "int64"}, "currency": {"type": "string"}, "duration": {"type": "string"}, "operation_id": {"type": "string"}}},
      "ScenarioExecution": {"type": "object", "required": ["scenario_id", "provider", "deterministic_configuration", "current_virtual_time", "payments", "async_operations"], "properties": {"scenario_id": {"type": "string"}, "provider": {"type": "object", "properties": {"id": {"type": "string"}, "profile": {"type": "string"}}}, "deterministic_configuration": {"type": "object", "properties": {"seed": {"type": "integer", "format": "int64"}}}, "current_virtual_time": {"type": "string", "format": "date-time"}, "payments": {"type": "array", "items": {"$ref": "#/components/schemas/Payment"}}, "async_operations": {"type": "array", "items": {"type": "object"}}}}
    },
    "responses": {
...
      "Scenario": {"description": "Scenario inputs.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Scenario"}}}},
      "ScenarioExecution": {"description": "Deterministic scenario execution result.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScenarioExecution"}}}},
      "InvalidRequest": {"description": "Invalid request.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
...
      "Unauthorized": {"description": "A valid admin bearer token is required.", "headers": {"WWW-Authenticate": {"schema": {"type": "string"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
      "NotFound": {"description": "Requested resource was not found.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
      "Conflict": {"description": "Operation rejected by current state or data.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
    }
  }
}